dbt-model-diff/pyproject.toml
[project]
name = "dbt-model-diff"
version = "0.5.0"
description = "OSS tool to compare dbt model data between git refs (Postgres + Redshift)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "rich>=13.0.0",
  "pyyaml>=6.0.0",
  "psycopg2-binary>=2.9.9"
]

[project.scripts]
dbt-model-diff = "dbt_model_diff.cli:main"
